// src/types/requests.ts
export interface ENSProfile {
  name?: string
  avatar?: string
  display?: string
  header?: string
  contenthash?: string
  records?: {
    avatar: string
    [key: string]: string
  }
  chains?: { [key: string]: string }
  fresh?: number
  resolver?: string
  errors?: { [key: string]: string }
}
// ...
export type FollowSortType = 'latest first' | 'earliest first' | 'follower count'
// ...
export interface FollowerResponse {
  address: Address
  ens: ENSProfile
  tags: string[]
  is_muted: boolean
  is_blocked: boolean
  is_following: boolean
}

export interface LatestFollowersResponse {
  address: Address
  efp_list_nft_token_id: string
  updated_at: string
}

export interface FollowingResponse {
  version: 1
  record_type: 'address' & string
  address: Address
  tags: string[]
  ens?: ENSProfile
}
// ...
export interface StatsResponse {
  followers_count: number
  following_count: number
}

export interface ProfileResponse {
  address: Address
  ens: ENSProfile
  fresh?: number
  resolver?: string
  primary_list?: string | null
  stats?: StatsResponse | undefined
  followers?: FollowerResponse[]
  following?: FollowingResponse[]
  chains?: Record<string, string>
  errors?: Record<string, unknown>
}

export interface ProfileDetailsResponse {
  address: Address
  ens: ENSProfile
  primary_list?: string | null
  ranks: {
    mutuals_rank: number
    followers_rank: number
    following_rank: number
    blocks_rank: number
  }
  stats?: StatsResponse | undefined
}
// ...
export type TagCountType = {
  tag: string
  count: number
}

export interface FollowingTagsResponse {
  token_id: string | number
  tags: string[]
  tagCounts: TagCountType[]
  taggedAddresses: {
    address: Address
    tag: string
  }[]
}

export type CommonFollower = {
  address: Address
  name: string | null
  avatar: string | null
  mutuals_rank: string
}

export interface CommonFollowersResponse {
  results: CommonFollower[]
  length: number
}

export interface ProfileListsResponse {
  primary_list?: string | null
  lists?: string[]
}

export type ProfileRoles = {
  isOwner: boolean
  isManager: boolean
  isUser: boolean
  listChainId: number
  listRecordsContract: Address
  listSlot: bigint
}